requirements.txt
tensorflow==2.15.0
streamlit==1.32.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.2
